refactor: new todo.js
