+++ src/index.js
@@ -1,5 +1,6 @@
 import { updateBackground } from './background.js';
-import { displayTodos, showGreeting } from './greeting.js';
+import { showGreeting } from './greeting.js';
+import { displayTodos } from './todo.js';
 import { paintQuote } from './quote.js';
 import { updateTime } from './time.js';
 import { displayWeather } from './weather.js';
